Move FilterConditionModel to shared module
