+++ projects/shared/src/lib/components/attribute-filter/attribute-filter.component.ts
@@ -5,7 +5,7 @@ import { debounceTime, take, takeUntil, tap } from 'rxjs/operators';
 import { DateTime } from 'luxon';
 import { AttributeType } from '@tailormap-viewer/api';
 import { FilterConditionEnum } from '@tailormap-viewer/api';
-import { FilterConditionModel } from '../../../../../core/src/lib/filter/models/filter-condition.model';
+import { FilterConditionModel } from '../../models/filter-condition.model';
 import { AttributeFilterHelper } from '../../helpers/attribute-filter.helper';
 import { FilterData, InputFilterData, OutputFilterData } from '../../models/attribute-filter-data.model';
